requirements.txt
numpy
pandas
scikit-learn
xgboost
skl2onnx
onnxmltools
onnx
pydantic
fastapi[standard]
onnxruntime
websockets
uvicorn
httpx
python-dotenv